[test]: Add Class object in each test script file (#992)
This is to get a better JUnitXML file and align all
the tests under the same level of hierarchy.
