+++ tests/test_portchannel.py
@@ -3,56 +3,57 @@ import time
 import re
 import json
 
-def test_PortChannel(dvs, testlog):
+class TestPortchannel(object):
+    def test_Portchannel(self, dvs, testlog):
 
-    # create port channel
-    db = swsscommon.DBConnector(0, dvs.redis_sock, 0)
-    ps = swsscommon.ProducerStateTable(db, "LAG_TABLE")
-    fvs = swsscommon.FieldValuePairs([("admin", "up"), ("mtu", "1500")])
+        # create port channel
+        db = swsscommon.DBConnector(0, dvs.redis_sock, 0)
+        ps = swsscommon.ProducerStateTable(db, "LAG_TABLE")
+        fvs = swsscommon.FieldValuePairs([("admin", "up"), ("mtu", "1500")])
 
-    ps.set("PortChannel0001", fvs)
+        ps.set("PortChannel0001", fvs)
 
-    # create port channel member
-    ps = swsscommon.ProducerStateTable(db, "LAG_MEMBER_TABLE")
-    fvs = swsscommon.FieldValuePairs([("status", "enabled")])
+        # create port channel member
+        ps = swsscommon.ProducerStateTable(db, "LAG_MEMBER_TABLE")
+        fvs = swsscommon.FieldValuePairs([("status", "enabled")])
 
-    ps.set("PortChannel0001:Ethernet0", fvs)
+        ps.set("PortChannel0001:Ethernet0", fvs)
 
-    time.sleep(1)
+        time.sleep(1)
 
-    # check asic db
-    asicdb = swsscommon.DBConnector(1, dvs.redis_sock, 0)
+        # check asic db
+        asicdb = swsscommon.DBConnector(1, dvs.redis_sock, 0)
 
-    lagtbl = swsscommon.Table(asicdb, "ASIC_STATE:SAI_OBJECT_TYPE_LAG")
-    lags = lagtbl.getKeys()
-    assert len(lags) == 1
+        lagtbl = swsscommon.Table(asicdb, "ASIC_STATE:SAI_OBJECT_TYPE_LAG")
+        lags = lagtbl.getKeys()
+        assert len(lags) == 1
 
-    lagmtbl = swsscommon.Table(asicdb, "ASIC_STATE:SAI_OBJECT_TYPE_LAG_MEMBER")
-    lagms = lagmtbl.getKeys()
-    assert len(lagms) == 1
+        lagmtbl = swsscommon.Table(asicdb, "ASIC_STATE:SAI_OBJECT_TYPE_LAG_MEMBER")
+        lagms = lagmtbl.getKeys()
+        assert len(lagms) == 1
 
-    (status, fvs) = lagmtbl.get(lagms[0])
-    for fv in fvs:
-        if fv[0] == "SAI_LAG_MEMBER_ATTR_LAG_ID":
-            assert fv[1] == lags[0]
-        elif fv[0] == "SAI_LAG_MEMBER_ATTR_PORT_ID":
-            assert dvs.asicdb.portoidmap[fv[1]] == "Ethernet0"
-        else:
-            assert False
+        (status, fvs) = lagmtbl.get(lagms[0])
+        for fv in fvs:
+            if fv[0] == "SAI_LAG_MEMBER_ATTR_LAG_ID":
+                assert fv[1] == lags[0]
+            elif fv[0] == "SAI_LAG_MEMBER_ATTR_PORT_ID":
+                assert dvs.asicdb.portoidmap[fv[1]] == "Ethernet0"
+            else:
+                assert False
 
-    # remove port channel member
-    ps = swsscommon.ProducerStateTable(db, "LAG_MEMBER_TABLE")
-    ps._del("PortChannel0001:Ethernet0")
+        # remove port channel member
+        ps = swsscommon.ProducerStateTable(db, "LAG_MEMBER_TABLE")
+        ps._del("PortChannel0001:Ethernet0")
 
-    # remove port channel
-    ps = swsscommon.ProducerStateTable(db, "LAG_TABLE")
-    ps._del("PortChannel0001")
+        # remove port channel
+        ps = swsscommon.ProducerStateTable(db, "LAG_TABLE")
+        ps._del("PortChannel0001")
 
-    time.sleep(1)
+        time.sleep(1)
 
-    # check asic db
-    lags = lagtbl.getKeys()
-    assert len(lags) == 0
+        # check asic db
+        lags = lagtbl.getKeys()
+        assert len(lags) == 0
 
-    lagms = lagmtbl.getKeys()
-    assert len(lagms) == 0
+        lagms = lagmtbl.getKeys()
+        assert len(lagms) == 0
