+++ tests/test_setro.py
@@ -4,37 +4,38 @@ import json
 import redis
 from pprint import pprint
 
-def test_SetReadOnlyAttribute(dvs, testlog):
+class TestSetRo(object):
+    def test_SetReadOnlyAttribute(self, dvs, testlog):
 
-    db = swsscommon.DBConnector(swsscommon.ASIC_DB, dvs.redis_sock, 0)
+        db = swsscommon.DBConnector(swsscommon.ASIC_DB, dvs.redis_sock, 0)
 
-    tbl = swsscommon.Table(db, "ASIC_STATE:SAI_OBJECT_TYPE_SWITCH")
+        tbl = swsscommon.Table(db, "ASIC_STATE:SAI_OBJECT_TYPE_SWITCH")
 
-    keys = tbl.getKeys()
+        keys = tbl.getKeys()
 
-    assert len(keys) == 1
+        assert len(keys) == 1
 
-    swVid = keys[0]
- 
-    r = redis.Redis(unix_socket_path=dvs.redis_sock, db=swsscommon.ASIC_DB)
+        swVid = keys[0]
 
-    swRid = r.hget("VIDTORID", swVid)
+        r = redis.Redis(unix_socket_path=dvs.redis_sock, db=swsscommon.ASIC_DB)
 
-    assert swRid is not None
+        swRid = r.hget("VIDTORID", swVid)
 
-    ntf = swsscommon.NotificationProducer(db, "SAI_VS_UNITTEST_CHANNEL")
+        assert swRid is not None
 
-    fvp = swsscommon.FieldValuePairs()
+        ntf = swsscommon.NotificationProducer(db, "SAI_VS_UNITTEST_CHANNEL")
 
-    ntf.send("enable_unittests", "true", fvp)
+        fvp = swsscommon.FieldValuePairs()
 
-    fvp = swsscommon.FieldValuePairs([('SAI_SWITCH_ATTR_PORT_MAX_MTU', '42')])
+        ntf.send("enable_unittests", "true", fvp)
 
-    key = "SAI_OBJECT_TYPE_SWITCH:" + swRid
+        fvp = swsscommon.FieldValuePairs([('SAI_SWITCH_ATTR_PORT_MAX_MTU', '42')])
 
-    print key
+        key = "SAI_OBJECT_TYPE_SWITCH:" + swRid
 
-    ntf.send("set_ro", key, fvp)
+        print key
 
-    # make action on appdb so orchagent will get RO value
-    # read asic db to see if orchagent behaved correctly
+        ntf.send("set_ro", key, fvp)
+
+        # make action on appdb so orchagent will get RO value
+        # read asic db to see if orchagent behaved correctly
